Rewrite Chip
