+++ packages/bento-design-system/src/index.ts
@@ -15,7 +15,7 @@ export * from "./Button/Button";
 export * from "./Button/ButtonLink";
 export * from "./Card/Card";
 export * from "./Checkbox/Checkbox";
-export * from "./Chip/createChip";
+export * from "./Chip/Chip";
 export * from "./ContentBlock/ContentBlock";
 export * from "./ContentWithSidebar/createContentWithSidebar";
 export * from "./DateField/createDateField";
