+++ packages/bento-design-system/src/util/ConfigurableTypes.ts
@@ -4,6 +4,7 @@ import { Object } from "ts-toolbelt";
 interface ConfigurableTypes {
   LocalizedString: string;
   SprinklesFn: typeof bentoSprinkles;
+  ChipCustomColors: never;
 }
 
 /**
